# src/backend/tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "BoostRequest": {
        "properties": {
          "level": {
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Level",
            "type": "integer"
          }
        },
        "required": [
          "level"
        ],
        "title": "BoostRequest",
        "type": "object"
      },
      "BoostResponse": {
        "properties": {
          "content_id": {
            "title": "Content Id",
            "type": "integer"
          },
          "editorial_boost": {
            "title": "Editorial Boost",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "content_id",
          "editorial_boost",
          "message"
        ],
        "title": "BoostResponse",
        "type": "object"
      },
...
      "ContentItemDetail": {
        "properties": {
          "added_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Added At"
          },
          "added_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Added By"
          },
          "ai_processed": {
            "default": false,
            "title": "Ai Processed",
            "type": "boolean"
          },
          "canonical_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Key"
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "dedupe_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dedupe Key"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "diversity_boost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diversity Boost"
          },
          "editorial_boost": {
            "default": 0,
            "title": "Editorial Boost",
            "type": "integer"
          },
          "entities": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entities"
          },
          "freshness_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness Tier"
          },
          "global_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Score"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "last_modified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Modified At"
          },
          "last_modified_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Modified By"
          },
          "manual_added": {
            "default": false,
            "title": "Manual Added",
            "type": "boolean"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "recency_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recency Score"
          },
          "recent_actions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EditorialActionRecord"
            },
            "title": "Recent Actions",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "suppressed": {
            "default": false,
            "title": "Suppressed",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topics": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topics"
          },
          "trend_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trend Score"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          }
        },
        "required": [
          "id",
          "title",
          "source",
          "source_url",
          "content_type"
        ],
        "title": "ContentItemDetail",
        "type": "object"
      },
      "ContentItemSummary": {
        "description": "Lightweight representation for list endpoints.",
        "properties": {
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "editorial_boost": {
            "default": 0,
            "title": "Editorial Boost",
            "type": "integer"
          },
          "freshness_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness Tier"
          },
          "global_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Score"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "manual_added": {
            "default": false,
            "title": "Manual Added",
            "type": "boolean"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "suppressed": {
            "default": false,
            "title": "Suppressed",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "source",
          "source_url",
          "content_type"
        ],
        "title": "ContentItemSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedContentResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContentItemSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PaginatedContentResponse",
        "type": "object"
      },
...
      "SubmitURLRequest": {
        "properties": {
          "importance_level": {
            "default": 0,
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Importance Level",
            "type": "integer"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 5,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "SubmitURLRequest",
        "type": "object"
      },
      "SubmitURLResponse": {
        "properties": {
          "content_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          },
          "duplicate": {
            "default": false,
            "title": "Duplicate",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "SubmitURLResponse",
        "type": "object"
      },
      "SuppressResponse": {
        "properties": {
          "content_id": {
            "title": "Content Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "suppressed": {
            "title": "Suppressed",
            "type": "boolean"
          }
        },
        "required": [
          "content_id",
          "suppressed",
          "message"
        ],
        "title": "SuppressResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-powered tech news aggregator API. Fetches articles from RSS feeds and videos from YouTube, processes with AI summarization, and serves personalized feeds.",
    "title": "blips-ai-news",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/editorial/content": {
      "get": {
        "description": "List content with editorial filters.",
        "operationId": "list_content_api_v1_admin_editorial_content_get",
        "parameters": [
          {
            "description": "UTC date YYYY-MM-DD",
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "UTC date YYYY-MM-DD",
              "title": "Day"
            }
          },
          {
            "description": "ARTICLE|VIDEO|REEL",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ARTICLE|VIDEO|REEL",
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "suppressed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Suppressed"
            }
          },
          {
            "in": "query",
            "name": "manual_added",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Manual Added"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "published_at",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Content",
        "tags": [
          "admin-editorial",
          "admin-editorial"
        ]
      }
    },
    "/api/v1/admin/editorial/content/submit": {
      "post": {
        "description": "Manually submit a URL for ingestion.\n\nRuns through the same canonicalization / dedupe logic as the\nautomatic ingestion pipeline.",
        "operationId": "submit_url_api_v1_admin_editorial_content_submit_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitURLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitURLResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Url",
        "tags": [
          "admin-editorial",
          "admin-editorial"
        ]
      }
    },
    "/api/v1/admin/editorial/content/{content_id}": {
      "get": {
        "description": "Get full content details + recent editorial actions.",
        "operationId": "get_content_detail_api_v1_admin_editorial_content__content_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "title": "Content Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentItemDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Content Detail",
        "tags": [
          "admin-editorial",
          "admin-editorial"
        ]
      }
    },
    "/api/v1/admin/editorial/content/{content_id}/boost": {
      "post": {
        "description": "Set editorial boost level (0-3).",
        "operationId": "boost_content_api_v1_admin_editorial_content__content_id__boost_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "title": "Content Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Boost Content",
        "tags": [
          "admin-editorial",
          "admin-editorial"
        ]
      }
    },
    "/api/v1/admin/editorial/content/{content_id}/suppress": {
      "post": {
        "description": "Suppress content (soft-delete). Removes from public feeds.",
        "operationId": "suppress_content_api_v1_admin_editorial_content__content_id__suppress_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "title": "Content Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuppressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suppress Content",
        "tags": [
          "admin-editorial",
          "admin-editorial"
        ]
      }
    },
    "/api/v1/admin/editorial/content/{content_id}/unsuppress": {
      "post": {
        "description": "Unsuppress content. Restores to public feeds.",
        "operationId": "unsuppress_content_api_v1_admin_editorial_content__content_id__unsuppress_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "title": "Content Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Key",
            "required": false,
            "schema": {
              "title": "X-Admin-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuppressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unsuppress Content",
        "tags": [
          "admin-editorial",
          "admin-editorial"
        ]
      }
    },
...
  }
}